v1.0.4 : fix minui_bm Android.mk, code cleanup and new adb modes
change usb tools labels and update scripts
Bootmenu full version now require BUILD_BOOTMENU_STANDALONE

+++ script/recovery_stable.sh
@@ -1,52 +1,111 @@
 #!/sbin/sh
 
-######## BootMenu Script v0.8.6
-######## Execute [Custom Recovery] Menu
+######## BootMenu Script
+######## Execute [Stable Recovery] Menu
 
 
 export PATH=/sbin:/system/xbin:/system/bin
 
 ######## Main Script
 
-# Moto 2.3.3 /tmp is a link to /data/tmp, bad thing ! &&
+## /tmp folder can be a link to /data/tmp, bad thing !
 [ -L /tmp ] && rm /tmp
-[ -L /etc ] && rm /etc
+mkdir -p /tmp
 
-mkdir /tmp
+mkdir -p /pds
+mkdir -p /res
+
+rm -f /etc
 mkdir /etc
-mkdir /res
 
 # hijack mke2fs & tune2fs CWM3
 rm -f /sbin/mke2fs
 rm -f /sbin/tune2fs
 rm -f /sbin/e2fsck
 
+rm -f /sdcard
+mkdir /sdcard
+
+chmod 755 /sbin
+chmod 755 /res
+
 cp -r -f /system/bootmenu/recovery/res/* /res/
-cp -r -f /system/bootmenu/recovery/sbin/* /sbin/
-chmod 755 /sbin/*
-cp /system/bootmenu/recovery/recovery.fstab /etc/recovery.fstab
+cp -p -f /system/bootmenu/recovery/sbin/* /sbin/
+cp -p -f /system/bootmenu/script/recoveryexit.sh /sbin/
 
-## adbd start
+chmod +rx /sbin/*
 
-rm -f /sbin/adbd
+rm -f /sbin/postrecoveryboot.sh
+
+if [ ! -e /etc/recovery.fstab ]; then
+  cp /system/bootmenu/recovery/recovery.fstab /etc/recovery.fstab
+fi
+
+mkdir -p /cache/recovery
+touch /cache/recovery/command
+touch /cache/recovery/log
+touch /cache/recovery/last_log
+touch /tmp/recovery.log
+
+killall adbd
+sleep 1
 ps | grep -v grep | grep adbd
 ret=$?
 
 if [ ! $ret -eq 0 ]; then
-  chmod 755 /system/bootmenu/script/adbd.sh
-  /system/bootmenu/script/adbd.sh
+   # chmod 755 /system/bootmenu/script/adbd.sh
+   # /system/bootmenu/script/adbd.sh
+
+   # don't use adbd here, will load many android process which locks /system
+   killall adbd
+   killall adbd.root
 fi
 
-
 #############################
-## mount in /sbin/postrecoveryboot.sh
+# mount in /sbin/postrecoveryboot.sh
 umount /system
-umount /data
+
+usleep 50000
+mount -t ext3 -o rw,noatime,nodiratime /dev/block/mmcblk1p21 /system
+
+# retry without type & options if not mounted
+[ ! -f /system/build.prop ] && mount -o rw /dev/block/mmcblk1p21 /system
+
+# set red led if problem with system, green led else
+
+echo 0 > /sys/class/leds/red/brightness
+echo 0 > /sys/class/leds/green/brightness
+echo 0 > /sys/class/leds/blue/brightness
+[ ! -f /system/build.prop ] && echo 1 > /sys/class/leds/red/brightness
+# [ -f /system/build.prop ] && echo 1 > /sys/class/leds/green/brightness
+
 #############################
 
 # turn on button backlight (back button is used in CWM Recovery 3.x)
 echo 1 > /sys/class/leds/button-backlight/brightness
 
+
 /sbin/recovery_stable
+
+
+# Post Recovery (back to bootmenu)
+
+# bootmenu doesnt support buttons
+echo 0 > /sys/class/leds/button-backlight/brightness
+
+# remount system & data if unmounted
+[ ! -d /data/data ] &&         mount -t ext3 -o rw,noatime,nodiratime,errors=continue /dev/block/userdata /data
+[ ! -f /system/build.prop ] && mount -t ext3 -o rw,noatime,nodiratime,errors=continue /dev/block/system /system
+
+if [ -f /system/build.prop ] ; then
+	echo 0 > /sys/class/leds/red/brightness
+	echo 0 > /sys/class/leds/green/brightness
+	echo 1 > /sys/class/leds/blue/brightness
+else
+	echo 1 > /sys/class/leds/red/brightness
+	echo 0 > /sys/class/leds/green/brightness
+	echo 0 > /sys/class/leds/blue/brightness
+fi
+
 
 exit
